workflow/requirements.txt
arrow==0.15.5
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-swagger==2.1.2
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-jsonapi==3.1.0
drf-json-api==0.1.1
drf-yasg==1.17.1
factory-boy==2.12.0
Faker==4.0.3
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pycodestyle==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
